(#653) Moved DefineTable and DefineTable<T> into the IOfflineStore.
